// src/frame-type.ts
/* Command Codes */
export enum FrameType {
  FAS_GetboardInfo = 0x01,
  FAS_GetMotorInfo = 0x05,
  FAS_SaveAllParameters = 0x10,
  FAS_GetRomParameter = 0x11,
  FAS_SetParameter = 0x12,
  FAS_GetParameter = 0x13,
  FAS_SetIOOutput = 0x20,
  FAS_SetIOInput = 0x21,
  FAS_GetIOInput = 0x22,
  FAS_GetIOOutput = 0x23,
  FAS_SetIOAssignMap = 0x24,
  FAS_GetIOAssignMap = 0x25,
  FAS_IOAssignMapReadROM = 0x26,
  FAS_TriggerOutput_RunA = 0x27,
  FAS_TriggerOutput_Status = 0x28,
  FAS_SetTriggerOutputEx = 0x7e,
  FAS_GetTriggerOutputEx = 0x7f,
  FAS_ServoEnable = 0x2a,
  FAS_ServoAlarmReset = 0x2b,
  FAS_GetAlarmType = 0x2e,
  FAS_MoveStop = 0x31,
  FAS_EmergencyStop = 0x32,
  FAS_MoveOriginSingleAxis = 0x33,
  FAS_MoveSingleAxisAbsPos = 0x34,
  FAS_MoveSingleAxisIncPos = 0x35,
  FAS_MoveToLimit = 0x36,
  FAS_MoveVelocity = 0x37,
  FAS_PositionAbsOverride = 0x38,
  FAS_PositionIncOverride = 0x39,
  FAS_VelocityOverride = 0x3a,
  FAS_MoveSingleAxisAbsPosEx = 0x80,
  FAS_MoveSingleAxisIncPosEx = 0x81,
  FAS_MoveVelocityEx = 0x82,
  FAS_GetAxisStatus = 0x40,
  FAS_GetIOAxisStatus = 0x41,
  FAS_GetMotionStatus = 0x42,
  FAS_GetAllStatus = 0x43,
  FAS_SetCommandPos = 0x50,
  FAS_GetCommandPos = 0x51,
  FAS_SetActualPos = 0x52,
  FAS_GetActualPos = 0x53,
  FAS_GetPosError = 0x54,
  FAS_GetActualVel = 0x55,
  FAS_ClearPosition = 0x56,
  FAS_MovePause = 0x58,
  FAS_PosTableReadItem = 0x60,
  FAS_PosTableWriteItem = 0x61,
  FAS_PosTableReadROM = 0x62,
  FAS_PosTableWriteROM = 0x63,
  FAS_PosTableRunItem = 0x64,
  FAS_PosTableReadOneItem = 0x6a,
  FAS_PosTableWriteOneItem = 0x6b,
  FAS_MovePush = 0x78,
  FAS_GetPushStatus = 0x79
}